Refactor: Store pricing table in a more structured format
This was for fun, I just wanted to see how to do it in rust. So
hard... it was so hard.

---

The following are for myself. It's a note about my Rust learning.
There is nowhere else to put it, so this is the place.

Learned a lot about Option and how to escape it. The error/null
handling in Rust is very strict and flexible at the same time. There
are so many ways I can platformed the error handling--I can be very
expressive about it or be very explicit, concise, or wordy, or
anything that fits the situation/intention. This feels very Ruby to
me. I'm really glad I picked Rust. 🥹 I want to explore more.

Oh, and the match thing is really powerful. Just wow! I never knew
that pattern matching could be this creative before. It also
compile error when I don't implement all cases. These things are...
gold, pure gold. I thought pattern matching was only for data, not on
the language feature itself (I don't know what to call it).