Add caching, debouncing or whatever they may call to make it possible to run inside the tmux status bar.
This program, from the start, was intended to be used inside the tmux
plugin or whatever shell scripts that we can mix it with it. Let's say
tmux status bar for example--we can't control how frequently it
refreshes its view. To avoid triggering api bans while keeping the
program light, it has to have some sort of denouncing to keep the fast
invoke rate away from the external api that the program may have just
blindingly hammered to.

That's the reason for this commit, Actually, I just want to have myself
relax by learning rust by writing it by hand, but turns out it wasn't
relaxing at all, lolololol.