Add terminal detection for better ergonomics.
Add code to autodetect the environment it runs in and selectively
render the spinning animation while waiting for fetching to finish.

This improves the ux. Users don't have to mandatory suffixing
commands with `--no-animate` everytime they do piping thing, for example

meter --no-animate | uplot --abc

It's even more wordy when in tmux status bar
[fg=#xx]#[bg=#yy]#(meter --no-animate) #(echo $(meter --no-animate --no-format) | \
... awk -xxx

It's okay for me personally but I don't know about others, so, why not?
It's harmless, I think.

Why do we need `--no-animate`?

There was a technical limitation before, I'm not certain about
the specific why, but we can't render spinning animations inside tmux
status bar. It will capture everything happens in the process it spawned
which will looks like this...

Retrieving...
Retrieving...
$2.2

And the tmux only show the first line because that's the only
space it has.

tmux will only shows this
Retrieving...

And the piping will break.

The flag `--no-animate` was added to fixes this, and the program will
render the spinner accordingly to this flag.