Update raw subcommand implementation.
It used to have two identical functions for remote fetching: one
returned raw string of body to user, another one sent the mapped struct
data to do a calculation before returning it to the user. From now on,
no more separate function. User will get the serialized version of
`Vec<UsageDataBucket>` as a return value which is the same data as what
calculation function will get when user request an aggregated data.

Note that it's not actually 'raw' anymore, despite the name.
Sorry that I can't think of a better name than this, 'inspect', 'export',
or 'dump' don't convey the same meaning as 'raw' does to me. It strongly
suggests that you can build something on top of it, a fine-grained script,
a dashboard, or whatever.

I will fix this somehow, maybe the laziest way is to note this in the
command --help?