Add custom Error enums, to use miette's features
Following up on the last commit, this introduces proper use of the Error
enum. Note that this isn't about error handling, but rather a
user-friendly report message.

To be honest, I just wanted to try out miette. 😅

This replaces simple error wrapping with miette features:
- detailed help messages.
- specific error codes.
- etc.

It makes the CLI output look much nicer and more helpful for the user.
Great!